[core][Android] Reduce the number of global references to JSIContext
